Non-zero base fixes. Fix vgaemu to work with it.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1923 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/mapping.h
@@ -107,8 +107,8 @@ void mapping_init(void);
 void mapping_close(void);
 
 void map_hardware_ram(void);
-int register_hardware_ram(int type, size_t base, size_t size);
-void *get_hardware_ram(size_t addr);
+int register_hardware_ram(int type, unsigned base, unsigned size);
+void *get_hardware_ram(unsigned addr);
 void list_hardware_ram(void (*print)(char *, ...));
 
 #endif /* _MAPPING_H_ */
